perf: split editor from hosted entry
