+++ apps/web/src/app.tsx
@@ -1,296 +1,8 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-	advanceDecisionView,
-	countUnanswered,
-	cursor,
-	Decisions,
-	Face,
-	PlanEditor,
-	QuestionnaireStore,
-	ThreadStore,
-	useHasPlanContent,
-	useQuestionnaires,
-	visibleDecisionView,
-} from "@chopin/editor";
+import { useEffect, useState } from "react";
 
-import { Chat } from "./chat/chat";
-import { decisionAttention, DecisionViewControl } from "./decision-view-control";
-import { DocumentPicker } from "./document-picker";
 import * as Api from "./api";
 import { HostedApp, HostedFailure, HostedLoading, HostedLogin } from "./hosted";
-import { peopleHere } from "./presence";
-import { RepositoryPicker } from "./repository-picker";
 import { clearRepositoryCache } from "./repository-cache";
-import { Wire } from "./wire";
-import { HEADING, useWorkspaceMode, useWorkspaceState, Workspace } from "./workspace";
-import { presentWorkspace } from "./workspace-model";
-
-import type { Session } from "@chopin/protocol";
-import type { DecisionView, DecisionViewState } from "@chopin/editor";
-import type { Status } from "./wire";
-
-function Header(
-	{
-		canEdit,
-		members,
-		label,
-		repository,
-		room,
-		userId,
-	}: {
-		canEdit: boolean;
-		members: Session.Member[];
-		label: string;
-		repository: Api.Repository;
-		room: string;
-		userId: string;
-	},
-) {
-	let people = peopleHere(members);
-	return (
-		<header className="room-header hairline-b relative flex min-h-12 shrink-0 flex-nowrap items-center gap-2 px-2 py-2 sm:h-12 sm:gap-3 sm:px-4 sm:py-0">
-			<a className="hidden text-sm font-semibold sm:inline" href="/">chopin</a>
-			<span aria-hidden="true" className="hidden h-4 hairline-l sm:block" />
-			<RepositoryPicker compact current={repository} key={userId} userId={userId} />
-			<div className="flex min-w-0 flex-1 items-center gap-2">
-				<span aria-hidden="true" className="hidden text-sm text-text-tertiary sm:inline">/</span>
-				<DocumentPicker
-					canEdit={canEdit}
-					current={{ id: room, title: label }}
-					repository={repository}
-					userId={userId}
-				/>
-			</div>
-			<div
-				aria-label={`People here: ${people.join(", ")}`}
-				className="room-members ml-auto flex shrink-0 items-center"
-				role="group"
-			>
-				{people.map(handle => (
-					<span className="room-member-face -ml-1.5 first:ml-0" key={handle.toLowerCase()}>
-						<Face handle={handle} ring size={24} />
-					</span>
-				))}
-				{people.length > 3 && (
-					<span
-						aria-hidden="true"
-						className="room-member-overflow ml-1 hidden text-sm text-text-tertiary"
-					>
-						+{people.length - 3}
-					</span>
-				)}
-			</div>
-		</header>
-	);
-}
-
-export function RoomWorkspace(
-	{
-		agent = true,
-		canEdit = true,
-		handle,
-		label,
-		repository,
-		room,
-		userId,
-	}: {
-		agent?: boolean;
-		canEdit?: boolean;
-		handle: string;
-		label: string;
-		repository: Api.Repository;
-		room: string;
-		userId: string;
-	},
-) {
-	let [wire, setWire] = useState<Wire>();
-	let [status, setStatus] = useState<Status>("connecting");
-	let [members, setMembers] = useState<Session.Member[]>([]);
-	let [effectiveCanEdit, setEffectiveCanEdit] = useState(canEdit);
-	let user = useMemo(() => cursor(handle), [handle]);
-	let mode = useWorkspaceMode();
-	let [workspace, dispatch] = useWorkspaceState();
-	let [questions] = useState(() => new QuestionnaireStore());
-	let [threads] = useState(() => new ThreadStore());
-	let [reveal, setReveal] = useState<{ widget: string; token: number }>();
-	let [planScrollTop, setPlanScrollTop] = useState(0);
-	let entries = useQuestionnaires(questions);
-	let unanswered = countUnanswered(entries);
-	let hasPlanContent = useHasPlanContent(questions);
-	let [decisionView, setDecisionView] = useState<DecisionViewState>(() => {
-		let stored = localStorage.getItem("chopin:view:document");
-		return {
-			phase: "initial",
-			preferred: stored === "decisions" ? "decisions" : "plan",
-		};
-	});
-	let view = visibleDecisionView(decisionView, hasPlanContent, unanswered);
-	let previousUnanswered = useRef(unanswered);
-	let [attention, setAttention] = useState(false);
-	let workspacePresentation = presentWorkspace(workspace, mode, view);
-	let conversationActive = workspacePresentation.conversationVisible;
-	let [conversationActivity, setConversationActivity] = useState({ unread: 0, busy: false });
-	let onConversationActivity = useCallback(
-		(event: { type: "message" | "working"; busy: boolean }) => {
-			setConversationActivity(current => ({
-				busy: event.busy,
-				unread: event.type === "message" && !conversationActive
-					? current.unread + 1
-					: current.unread,
-			}));
-		},
-		[conversationActive],
-	);
-
-	useEffect(() => {
-		setDecisionView(state => advanceDecisionView(state, hasPlanContent, unanswered));
-	}, [hasPlanContent, unanswered]);
-
-	useEffect(() => {
-		if (!conversationActive) return;
-		setConversationActivity(current => current.unread === 0 ? current : { ...current, unread: 0 });
-	}, [conversationActive]);
-
-	useEffect(() => {
-		let previous = previousUnanswered.current;
-		previousUnanswered.current = unanswered;
-		if (!decisionAttention(previous, unanswered)) return;
-		setAttention(true);
-		let timer = window.setTimeout(() => setAttention(false), 200);
-		return () => window.clearTimeout(timer);
-	}, [unanswered]);
-
-	let selectView = (next: DecisionView, revealFirst = true) => {
-		setDecisionView(state => ({ ...state, preferred: next }));
-		if (hasPlanContent) localStorage.setItem("chopin:view:document", next);
-		if (next === "decisions" && revealFirst) {
-			let first = entries.find(entry =>
-				entry.value.questions.some(question => question.answer === undefined)
-			);
-			setReveal({ widget: first?.id ?? "", token: Date.now() });
-		}
-	};
-
-	let selectDestination = (destination: "plan" | "decisions") => {
-		selectView(destination, mode === "split");
-		dispatch({ type: "set-conversation", open: false });
-	};
-
-	let setDesktopConversationOpen = (open: boolean) => {
-		dispatch({ type: "set-desktop-conversation", open });
-		if (!open) dispatch({ type: "set-conversation", open: false });
-	};
-
-	let showPlan = (widget: string, question: string) => {
-		selectDestination("plan");
-		requestAnimationFrame(() => {
-			questions.reveal(widget, question);
-			let card = document.querySelector<HTMLElement>(
-				`[data-document-view="plan"] article[data-plan-sidecar-questionnaire="${
-					CSS.escape(widget)
-				}"]`,
-			);
-			if (!card) return;
-			card.tabIndex = -1;
-			card.focus();
-		});
-	};
-
-	useEffect(() => {
-		setEffectiveCanEdit(canEdit);
-	}, [canEdit, room]);
-
-	useEffect(() => {
-		let socket = new Wire({
-			channelId: room,
-			onAuthenticationRequired: () => location.assign("/"),
-			onStatus: next => {
-				setStatus(next);
-			},
-		});
-		setWire(socket);
-
-		let off = [
-			socket.on<Session.Hello>("session:hello", frame => {
-				setMembers(frame.members);
-				setEffectiveCanEdit(frame.canEdit);
-			}),
-			socket.on<Session.Presence>("session:presence", frame => setMembers(frame.members)),
-			socket.on<Session.Access>("session:access", frame => setEffectiveCanEdit(frame.canEdit)),
-			threads.listen(socket),
-		];
-
-		return () => {
-			for (let unsubscribe of off) unsubscribe();
-			socket.dispose();
-			setWire(undefined);
-		};
-	}, [room, handle, threads]);
-
-	return (
-		<Workspace
-			chat={
-				<Chat
-					active={conversationActive}
-					agent={agent}
-					connected={status === "connected" && effectiveCanEdit}
-					handle={handle}
-					onActivity={onConversationActivity}
-					wire={wire}
-				/>
-			}
-			conversationActivity={conversationActivity}
-			header={
-				<Header
-					canEdit={effectiveCanEdit}
-					members={members}
-					label={label}
-					repository={repository}
-					room={room}
-					userId={userId}
-				/>
-			}
-			controls={
-				<DecisionViewControl
-					attention={attention}
-					onView={selectDestination}
-					unanswered={unanswered}
-					view={view}
-				/>
-			}
-			mode={mode}
-			onDesktopConversationOpen={setDesktopConversationOpen}
-			onConversationOpen={open => dispatch({ type: "set-conversation", open })}
-			onDestination={selectDestination}
-			decisions={
-				<Decisions
-					connected={status === "connected" && effectiveCanEdit}
-					headingId={HEADING.decisions}
-					onShowPlan={showPlan}
-					reveal={reveal}
-					store={questions}
-					wire={wire}
-				/>
-			}
-			plan={
-				<PlanEditor
-					commentPresentation={mode === "split" ? "popover" : "sheet"}
-					connection={status}
-					onScrollTop={setPlanScrollTop}
-					questions={questions}
-					readOnly={!effectiveCanEdit}
-					scrollTop={planScrollTop}
-					threads={threads}
-					user={user}
-					wire={wire}
-				/>
-			}
-			state={workspace}
-			unanswered={unanswered}
-			view={view}
-		/>
-	);
-}
 
 export function App() {
 	let [session, setSession] = useState<Api.Session>();
@@ -330,5 +42,5 @@ export function App() {
 	if (error) return <HostedFailure error={error} />;
 	if (!session || !repositoryCacheReady) return <HostedLoading />;
 	if (!session.user) return <HostedLogin />;
-	return <HostedApp Workspace={RoomWorkspace} agent={session.agent} user={session.user} />;
+	return <HostedApp agent={session.agent} user={session.user} />;
 }
